+++ NEWS.md
@@ -3,6 +3,7 @@
 ## Changes to functions
 
 * Better support for different model families in `sjp.glm` and `sjp.glmer`.
+* `sjt.lm`, `sjt.lmer`, `sjt.glm` and `sjt.glmer` get a `showDeviance` argument to display model's deviance in the table summary.
 
 ## Bug fixes